requirements.txt
selenium
pydantic
pandas
beautifulsoup4
lxml
pyyaml
